Break out config builder
